requirements.txt
streamlit
pyttsx3
pyinstaller
mysql-connector-python
